chore: add sha1 utility module
