+++ scripts/sha1.js
@@ -1,0 +1,12 @@
+const fs = require("fs");
+const crypto = require("crypto");
+
+module.exports = function sha1(file) {
+  return new Promise((resolve, reject) => {
+    const hash = crypto.createHash("sha1");
+    const stream = fs.createReadStream(file);
+    stream.on("error", reject);
+    stream.on("data", (chunk) => hash.update(chunk));
+    stream.on("end", () => resolve(hash.digest("hex")));
+  });
+};
